Extract CustomAssertionsTestStatus from test_cs_status
I'm going to want some of these same custom assertions in a new test
module for testing TestStatus directly, so I'm extracting this custom
assertions class into its own module (and similarly extracting the unit
tests of this custom assertions class into their own test module).
