+++ scripts/lib/CIME/tests/test_cs_status.py
@@ -9,137 +9,9 @@ import six
 import six_additions
 from CIME.cs_status import cs_status
 from CIME import test_status
+from CIME.tests.custom_assertions_test_status import CustomAssertionsTestStatus
 
-# ========================================================================
-# Custom assertions
-# ========================================================================
-
-class CustomAssertions(unittest.TestCase):
-
-    def assert_status_of_phase(self, output, status, phase, test_name):
-        """Asserts that 'output' contains a line showing the given
-        status for the given phase for the given test_name"""
-        expected = re.compile(r'^ *{} +'.format(re.escape(status)) +
-                              self._test_name_and_phase_regex(test_name, phase),
-                              flags=re.MULTILINE)
-
-        six.assertRegex(self, output, expected)
-
-    def assert_phase_absent(self, output, phase, test_name):
-        """Asserts that 'output' does not contain a status line for the
-        given phase and test_name"""
-        expected = re.compile(r'^.* +' +
-                              self._test_name_and_phase_regex(test_name, phase),
-                              flags=re.MULTILINE)
-
-        six_additions.assertNotRegex(self, output, expected)
-
-    def assert_core_phases(self, output, test_name, fails):
-        """Asserts that 'output' contains a line for each of the core test
-        phases for the given test_name. All results should be PASS
-        except those given by the fails list, which should be FAILS.
-        """
-        for phase in test_status.CORE_PHASES:
-            if phase in fails:
-                status = test_status.TEST_FAIL_STATUS
-            else:
-                status = test_status.TEST_PASS_STATUS
-            self.assert_status_of_phase(output=output,
-                                        status=status,
-                                        phase=phase,
-                                        test_name=test_name)
-
-    @staticmethod
-    def _test_name_and_phase_regex(test_name, phase):
-        """Returns a regex matching the portion of a TestStatus line
-        containing the test name and phase"""
-        # The main purpose of extracting this into a shared method is:
-        # assert_phase_absent could wrongly pass if the format of the
-        # TestStatus output changed without that method's regex
-        # changing. By making its regex shared as much as possible with
-        # the regex in assert_status_of_phase, we decrease the chances
-        # of these false passes.
-        return r'{} +{} *$'.format(re.escape(test_name), re.escape(phase))
-
-class TestCustomAssertions(CustomAssertions):
-
-    @staticmethod
-    def output_line(status, test_name, phase):
-        return status + ' ' + test_name + ' ' + phase + '\n'
-
-    # Note: No explicit tests of assert_status_of_phase: this is covered
-    # sufficiently via tests of assert_core_phases
-
-    def test_assert_phase_absent_passes(self):
-        """assert_phase_absent should pass when the phase is absent for
-        the given test_name"""
-        test_name1 = 'my.test.name1'
-        test_name2 = 'my.test.name2'
-        output = self.output_line('PASS', test_name1, 'PHASE1')
-        output += self.output_line('PASS', test_name2, 'PHASE2')
-
-        self.assert_phase_absent(output, 'PHASE2', test_name1)
-        self.assert_phase_absent(output, 'PHASE1', test_name2)
-
-    def test_assert_phase_absent_fails(self):
-        """assert_phase_absent should fail when the phase is present for
-        the given test_name"""
-        test_name = 'my.test.name'
-        output = self.output_line('PASS', test_name, 'PHASE1')
-
-        with self.assertRaises(AssertionError):
-            self.assert_phase_absent(output, 'PHASE1', test_name)
-
-    def test_assert_core_phases_passes(self):
-        """assert_core_phases passes when it should"""
-        output = ''
-        fails = [test_status.CORE_PHASES[1]]
-        test_name = 'my.test.name'
-        for phase in test_status.CORE_PHASES:
-            if phase in fails:
-                status = test_status.TEST_FAIL_STATUS
-            else:
-                status = test_status.TEST_PASS_STATUS
-            output = output + self.output_line(status, test_name, phase)
-
-        self.assert_core_phases(output, test_name, fails)
-
-    def test_assert_core_phases_missing_phase_fails(self):
-        """assert_core_phases fails if there is a missing phase"""
-        output = ''
-        test_name = 'my.test.name'
-        for phase in test_status.CORE_PHASES:
-            if phase != test_status.CORE_PHASES[1]:
-                output = output + self.output_line(test_status.TEST_PASS_STATUS, test_name, phase)
-
-        with self.assertRaises(AssertionError):
-            self.assert_core_phases(output, test_name, fails=[])
-
-    def test_assert_core_phases_wrong_status_fails(self):
-        """assert_core_phases fails if a phase has the wrong status"""
-        output =''
-        test_name = 'my.test.name'
-        for phase in test_status.CORE_PHASES:
-            output = output + self.output_line(test_status.TEST_PASS_STATUS, test_name, phase)
-
-        with self.assertRaises(AssertionError):
-            self.assert_core_phases(output, test_name, fails=[test_status.CORE_PHASES[1]])
-
-    def test_assert_core_phases_wrong_name_fails(self):
-        """assert_core_phases fails if the test name is wrong"""
-        output =''
-        test_name = 'my.test.name'
-        for phase in test_status.CORE_PHASES:
-            output = output + self.output_line(test_status.TEST_PASS_STATUS, test_name, phase)
-
-        with self.assertRaises(AssertionError):
-            self.assert_core_phases(output, 'my.test', fails=[])
-
-# ========================================================================
-# Actual test class
-# ========================================================================
-
-class TestCsStatus(CustomAssertions):
+class TestCsStatus(CustomAssertionsTestStatus):
 
     # ------------------------------------------------------------------------
     # Test helper functions
